Projeto Alura imersão I.A/requirements.txt
google-generativeai==0.4.1